+++ CHANGES.markdown
@@ -2,6 +2,7 @@
  - Visually distinguish error message from requirements in the summary (#100)
  - Export `formatException` from `Test.Hspec.Formatters`
  - Add `--fast-fail` (#82)
+ - Print a summary on UserInterrupt/ctrl-c (#107)
 
 ## Changes in 1.4.3
  - Add `--dry-run` (#111)